model exports an index
